# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Permissions service stores user resources and allows to manage access to resources",
    "title": "Permissions API"
  },
  "paths": {
...
    "/admin/volumes": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "summary": "Get all volumes (admin only).",
        "operationId": "GetAllVolumes",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/Filters"
          },
          {
            "$ref": "#/parameters/PageNum"
          },
          {
            "$ref": "#/parameters/PerPageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "volumes response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/VolumeWithPermissions"
              }
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "delete": {
        "tags": [
          "Volumes"
        ],
        "summary": "Delete all user volumes (admin only).",
        "operationId": "DeleteAllUserVolumes",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          }
        ],
        "responses": {
          "200": {
            "description": "volumes deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
    "/volumes": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "summary": "Get user volumes.",
        "operationId": "GetUserVolumes",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/Filters"
          }
        ],
        "responses": {
          "200": {
            "description": "volumes response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/VolumeWithPermissions"
              }
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "post": {
        "description": "Should be chosen first storage, where free space allows to create volume with provided capacity.",
        "tags": [
          "Volumes"
        ],
        "summary": "Create Volume for User by Tariff.",
        "operationId": "CreateVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VolumeCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "volume created"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/volumes/{label}": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "summary": "Get volume.",
        "operationId": "GetVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "type": "string",
            "name": "label",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "volume response",
            "schema": {
              "$ref": "#/definitions/VolumeWithPermissions"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "put": {
        "tags": [
          "Volumes"
        ],
        "summary": "Resize volume.",
        "operationId": "ResizeVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VolumeResizeRequest"
            }
          },
          {
            "type": "string",
            "name": "label",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "volume resized"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "delete": {
        "tags": [
          "Volumes"
        ],
        "summary": "Delete volume.",
        "operationId": "DeleteVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "type": "string",
            "name": "label",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "volume deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
...
    "/volumes/{label}/rename": {
      "put": {
        "tags": [
          "Volumes"
        ],
        "summary": "Rename volume.",
        "operationId": "RenameVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VolumeRenameRequest"
            }
          },
          {
            "type": "string",
            "name": "label",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "volume renamed"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    }
  },
  "definitions": {
    "AccessLevel": {
      "type": "string",
      "enum": [
        "none",
        "read",
        "readdelete",
        "write",
        "owner"
      ]
    },
...
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/permissions/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/permissions/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/permissions/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/permissions/vendor/github.com/containerum/cherry"
    },
...
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/permissions/vendor/github.com/containerum/cherry"
    },
...
    "Permission": {
      "description": "Permission represents information about user permission to resource",
      "type": "object",
      "properties": {
        "access": {
          "$ref": "#/definitions/AccessLevel"
        },
        "access_level_change_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "AccessLevelChangeTime"
        },
        "create_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreateTime"
        },
        "kind": {
          "$ref": "#/definitions/ResourceType"
        },
        "login": {
          "type": "string",
          "format": "email",
          "x-go-name": "UserLogin"
        },
        "new_access_level": {
          "$ref": "#/definitions/AccessLevel"
        },
        "perm_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "resource_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ResourceID"
        },
        "user_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "git.containerum.net/ch/permissions/pkg/model"
    },
...
    "ResourceType": {
      "type": "string",
      "x-go-package": "git.containerum.net/ch/permissions/pkg/model"
    },
...
    "VolumeCreateRequest": {
      "description": "VolumeCreateRequest is a request object for creating volume",
      "type": "object",
      "properties": {
        "label": {
          "type": "string",
          "x-go-name": "Label"
        },
        "tariff_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "TariffID"
        }
      },
      "x-go-package": "git.containerum.net/ch/permissions/pkg/model"
    },
    "VolumeRenameRequest": {
      "description": "VolumeRenameRequest is a request object for renaming volume",
      "type": "object",
      "properties": {
        "label": {
          "type": "string",
          "x-go-name": "Label"
        }
      },
      "x-go-package": "git.containerum.net/ch/permissions/pkg/model"
    },
    "VolumeResizeRequest": {
      "description": "VolumeResizeRequest contains parameters for changing volume size",
      "type": "object",
      "properties": {
        "tariff_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "TariffID"
        }
      },
      "x-go-package": "git.containerum.net/ch/permissions/pkg/model"
    },
...
    "VolumeWithPermissions": {
      "description": "VolumeWithPermissions is a response object for get requests",
      "type": "object",
      "properties": {
        "access": {
          "$ref": "#/definitions/AccessLevel"
        },
        "access_level_change_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "AccessLevelChangeTime"
        },
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "capacity": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Capacity"
        },
        "create_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreateTime"
        },
        "delete_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DeleteTime"
        },
        "deleted": {
          "type": "boolean",
          "x-go-name": "Deleted"
        },
        "gluster_name": {
          "type": "string",
          "x-go-name": "GlusterName"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "kind": {
          "$ref": "#/definitions/ResourceType"
        },
        "label": {
          "type": "string",
          "x-go-name": "Label"
        },
        "login": {
          "type": "string",
          "format": "email",
          "x-go-name": "UserLogin"
        },
        "namespace_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "NamespaceID"
        },
        "new_access_level": {
          "$ref": "#/definitions/AccessLevel"
        },
        "owner_user_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "OwnerUserID"
        },
        "perm_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "replicas": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Replicas"
        },
        "resource_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ResourceID"
        },
        "storage_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "StorageID"
        },
        "tariff_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "TariffID"
        },
        "user_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "UserID"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Permission"
          },
          "x-go-name": "Permissions"
        }
      },
      "x-go-name": "VolumeWithPermissionsJSON",
      "x-go-package": "git.containerum.net/ch/permissions/pkg/model"
    },
...
  },
  "parameters": {
    "Filters": {
      "type": "string",
      "description": "A set of filters separated with comma. See \"dao\" package for more information.",
      "name": "filter",
      "in": "query"
    },
    "PageNum": {
      "minimum": 0,
      "type": "integer",
      "name": "page",
      "in": "query",
      "required": true
    },
    "PerPageLimit": {
      "minimum": 0,
      "type": "integer",
      "name": "per_page",
      "in": "query",
      "required": true
    },
    "SubstitutedUserID": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/SubstitutedUserID"
    },
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}